refactor: coin entity 하위 패키지로  response 패키지 정의
