+++ app/src/main/java/com/coinkiri/coinkiri/data/coin/mapper/CoinMapper.kt
@@ -1,7 +1,7 @@
 package com.coinkiri.coinkiri.data.coin.mapper
 
 import com.coinkiri.coinkiri.data.coin.dto.response.CoinResponseDto
-import com.coinkiri.coinkiri.domain.coin.entity.CoinResponseEntity
+import com.coinkiri.coinkiri.domain.coin.entity.response.CoinResponseEntity
 
 fun CoinResponseDto.toCoinEntity(): CoinResponseEntity =
     CoinResponseEntity(
