+++ app/src/main/java/com/coinkiri/coinkiri/domain/coin/entity/response/CoinResponseEntity.kt
@@ -1,4 +1,4 @@
-package com.coinkiri.coinkiri.domain.coin.entity
+package com.coinkiri.coinkiri.domain.coin.entity.response
 
 data class CoinResponseEntity(
     val marketName: String,
