+++ app/src/main/java/com/coinkiri/coinkiri/domain/coin/repository/CoinRepository.kt
@@ -1,6 +1,6 @@
 package com.coinkiri.coinkiri.domain.coin.repository
 
-import com.coinkiri.coinkiri.domain.coin.entity.CoinResponseEntity
+import com.coinkiri.coinkiri.domain.coin.entity.response.CoinResponseEntity
 
 interface CoinRepository {
     suspend fun getCoinList(): Result<List<CoinResponseEntity>>
