+++ app/src/main/java/com/coinkiri/coinkiri/domain/coin/usecase/GetCoinListUseCase.kt
@@ -1,6 +1,6 @@
 package com.coinkiri.coinkiri.domain.coin.usecase
 
-import com.coinkiri.coinkiri.domain.coin.entity.CoinResponseEntity
+import com.coinkiri.coinkiri.domain.coin.entity.response.CoinResponseEntity
 import com.coinkiri.coinkiri.domain.coin.repository.CoinRepository
 import javax.inject.Inject
